Initial refactor and prep work
