+++ lib/components/activitycard.dart
@@ -299,27 +299,6 @@ class ActivityCard extends StatelessWidget {
                         ],
                       ),
                     ),
-                    // Row(
-                    //   mainAxisAlignment: MainAxisAlignment.start,
-                    //   children: [
-                    //     Icon(
-                    //       Icons.location_on,
-                    //       size: 16.0,
-                    //       color: Color(0xFFBE4C59),
-                    //     ),
-                    //     FittedBox(
-                    //       fit: BoxFit.fitWidth,
-                    //       child: Text(
-                    //         '  ' + zone,
-                    //         style: TextStyle(
-                    //           color: asoPrimary,
-                    //           fontWeight: FontWeight.w500,
-                    //           fontSize: 16.0,
-                    //         ),
-                    //       ),
-                    //     ),
-                    //   ],
-                    // ),
                   ],
                 ),
               ),
@@ -333,121 +312,5 @@ class ActivityCard extends StatelessWidget {
         ],
       ),
     );
-      // child: Card(
-      //   color: Colors.transparent,
-      //   shape: RoundedRectangleBorder( 
-      //     borderRadius: BorderRadius.circular(30),
-      //   ),
-      //   child: InkWell( 
-      //     splashColor: Colors.grey[200],
-      //     onTap: () { 
-      //       print('owo it works');
-      //       print(MediaQuery.of(context).size.height / 4);
-      //     },
-      //     child: Column(
-      //       mainAxisSize: MainAxisSize.min,
-      //       children: <Widget>[
-      //         Align(
-      //           alignment: Alignment.centerRight,
-      //           widthFactor: 0.5,
-      //           heightFactor: 0.5,
-      //           child: FittedBox(
-      //             child: Image.network(image, width: 200, height: 200),
-      //             fit: BoxFit.fitHeight
-      //           )
-      //         ),  
-      //         ListTile(
-      //           leading: Icon( 
-      //             Icons.location_on,
-      //             size: 32.0,
-      //             color: Color(0xFFBE4C59),
-      //           ),
-      //           title: Text(
-      //             title,
-      //             style: TextStyle( 
-      //               fontSize: 24.0,
-      //             )
-      //           ),
-      //           // remove placeholder subtitle later
-      //           subtitle: Text(desc + 'Check it out!'),
-      //         ),  
-      //         ButtonBar(
-      //           children: <Widget>[
-      //             pageButton
-      //           ],
-      //         ),
-      //       ],
-      //     ),
-      //   ),
-      // ),
-      // Stack(
-      //   children: <Widget>[
-      //     Positioned(
-      //       bottom: 0,
-      //       left: 0,
-      //       right: 0,
-      //       child: Container(
-      //         height: 221,
-      //         decoration: BoxDecoration(
-      //           color: Colors.white,
-      //           borderRadius: BorderRadius.circular(29),
-      //           boxShadow: [
-      //             BoxShadow(
-      //               offset: Offset(0, 10),
-      //               blurRadius: 33,
-      //               color: Colors.grey,
-      //             ),
-      //           ],
-      //         ),
-      //       ),
-      //     ),
-      //     Image.network(image, height:200,width: 150),
-      //     Positioned(
-      //       top: 35,
-      //       right: 10,
-      //       child: Column(
-      //         children: <Widget>[
-      //           IconButton(
-      //             icon: Icon(
-      //               Icons.favorite_border,
-      //             ),
-      //             onPressed: () {},
-      //           ),
-      //         ],
-      //       ),
-      //     ),
-      //     Positioned(
-      //       top: 160,
-      //       child: Container(
-      //         height: 85,
-      //         width: 202,
-      //         child: Column(
-      //           crossAxisAlignment: CrossAxisAlignment.start,
-      //           children: <Widget>[
-      //             Padding(
-      //               padding: EdgeInsets.only(left: 24),
-      //               child: RichText(
-      //                 maxLines: 2,
-      //                 text: TextSpan(
-      //                   style: TextStyle(color: Colors.black),
-      //                   children: [
-      //                     TextSpan(
-      //                       text: "$title\n",
-      //                       style: TextStyle(
-      //                         fontWeight: FontWeight.bold,
-      //                       ),
-      //                     ),
-      //                   ],
-      //                 ),
-      //               ),
-      //             ),
-      //             Spacer(),
-      //             pageButton
-      //           ],
-      //         ),
-      //       ),
-      //     ),
-      //   ],
-      // ),
   }
 }
