+++ lib/components/card.dart
@@ -86,76 +86,6 @@ class ArtListCard extends StatelessWidget {
             ),
           ],
         ),
-
-        // Stack(
-        //   children: <Widget>[
-        //     Positioned(
-        //       bottom: 0,
-        //       left: 0,
-        //       right: 0,
-        //       child: Container(
-        //         height: 221,
-        //         decoration: BoxDecoration(
-        //           color: Colors.white,
-        //           borderRadius: BorderRadius.circular(29),
-        //           boxShadow: [
-        //             BoxShadow(
-        //               offset: Offset(0, 10),
-        //               blurRadius: 33,
-        //               color: Colors.grey,
-        //             ),
-        //           ],
-        //         ),
-        //       ),
-        //     ),
-        //     Image.network(image, height:200,width: 150),
-        //     Positioned(
-        //       top: 35,
-        //       right: 10,
-        //       child: Column(
-        //         children: <Widget>[
-        //           IconButton(
-        //             icon: Icon(
-        //               Icons.favorite_border,
-        //             ),
-        //             onPressed: () {},
-        //           ),
-        //         ],
-        //       ),
-        //     ),
-        //     Positioned(
-        //       top: 160,
-        //       child: Container(
-        //         height: 85,
-        //         width: 202,
-        //         child: Column(
-        //           crossAxisAlignment: CrossAxisAlignment.start,
-        //           children: <Widget>[
-        //             Padding(
-        //               padding: EdgeInsets.only(left: 24),
-        //               child: RichText(
-        //                 maxLines: 2,
-        //                 text: TextSpan(
-        //                   style: TextStyle(color: Colors.black),
-        //                   children: [
-        //                     TextSpan(
-        //                       text: "$title\n",
-        //                       style: TextStyle(
-        //                         fontWeight: FontWeight.bold,
-        //                       ),
-        //                     ),
-        //                   ],
-        //                 ),
-        //               ),
-        //             ),
-        //             Spacer(),
-        //             pageButton
-        //           ],
-        //         ),
-        //       ),
-        //     ),
-        //   ],
-        // ),
       ),
     );
   }
