+++ lib/graphql/AnnouncementQueries.dart
@@ -1,10 +1,3 @@
-class Announcement {
-  String title;
-  String desc;
-
-  Announcement(this.title, this.desc);
-}
-
 class AnnouncementQueries {
   String getAll = """ 
     {
