+++ lib/graphql/GraphQLConfiguration.dart
