+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:artsideout_app/theme.dart';
 import 'package:artsideout_app/layout.dart';
 // GraphQL
 import 'package:graphql_flutter/graphql_flutter.dart';
-import "package:artsideout_app/graphql/config.dart";
+import "package:artsideout_app/graphql/GraphQLConfiguration.dart";
 
 GraphQLConfiguration graphQLConfiguration = GraphQLConfiguration();
