+++ lib/models/Announcement.dart
@@ -1,0 +1,6 @@
+class Announcement {
+  String title;
+  String desc;
+
+  Announcement(this.title, this.desc);
+}
