+++ lib/pages/profile/ProfileDetailPage.dart
@@ -1,7 +1,6 @@
 import 'package:artsideout_app/components/profile/ProfileDetailWidget.dart';
-import 'package:artsideout_app/theme.dart';
+import 'package:artsideout_app/models/Profile.dart';
 import 'package:flutter/material.dart';
-import 'package:artsideout_app/graphql/Profile.dart';
 
 class ProfileDetailPage extends StatefulWidget {
   final Profile profile;
